requirements.txt
flask>=3.1
google-genai>=1.0
pyexiftool>=0.5
click>=8.1
tqdm>=4.66
Pillow>=10.0
pywebview>=6.0
keyring>=24
reportlab>=4.0
